sorted channel sync
