+++ part-1-exercises/toDoTypes.go
